Произведена оптимизация под разные разрешения. Теперь все действия из логики следует вычислять как percentX * (data) или percentY * (data). Вылет за экран, движения по ОХ и ОУ, ширина игрока и кнопка переключения управления работают в процентном соотношении. Юрыч, твоё зацикленное движение по OY ломает концепцию апдейта, так как работает в отдельном независимом потоке.

+++ src/main/java/core/GameObject.java
@@ -24,7 +24,7 @@ abstract public class GameObject {
     /**
      * Конструктор
      */
-    public GameObject(BaseGameActivity activity, Engine engine, float positionX, float positionY, int spriteWidth, int spriteHeight) {
+    public GameObject(BaseGameActivity activity, Engine engine, float positionX, float positionY, float spriteWidth, float spriteHeight) {
         this.activity = activity;
         this.engine = engine;
         atlas = getNewAtlas();
